requirements.txt
bottle
jinja2
rpi.gpio
requests
